Add TSX support
