+++ lib/rouge/demos/tsx
@@ -1,0 +1,17 @@
+class HelloWorld extends React.Component<{date: Date}, void> {
+  render() {
+    return (
+      <p>
+        Hello, <input type="text" placeholder="Your name here" />!
+        It is {this.props.date.toTimeString()}
+      </p>
+    );
+  }
+}
+
+setInterval(function() {
+  ReactDOM.render(
+    <HelloWorld date={new Date()} />,
+    document.getElementById('example')
+  );
+}, 500);
