+++ lib/rouge/lexers/typescript.rb
@@ -3,8 +3,11 @@
 module Rouge
   module Lexers
     load_lexer 'javascript.rb'
+    load_lexer 'typescript/common.rb'
 
     class Typescript < Javascript
+      include TypescriptCommon
+
       title "TypeScript"
       desc "TypeScript, a superset of JavaScript"
 
@@ -14,32 +17,6 @@ module Rouge
       filenames '*.ts', '*.d.ts'
 
       mimetypes 'text/typescript'
-
-      def self.keywords
-        @keywords ||= super + Set.new(%w(
-          is namespace static private protected public
-          implements readonly
-        ))
-      end
-
-      def self.declarations
-        @declarations ||= super + Set.new(%w(
-          type abstract
-        ))
-      end
-
-      def self.reserved
-        @reserved ||= super + Set.new(%w(
-          string any void number namespace module
-          declare default interface keyof
-        ))
-      end
-
-      def self.builtins
-        @builtins ||= super + %w(
-          Pick Partial Readonly Record
-        )
-      end
     end
   end
 end
